[nfc] Split ASTGen into files + extensions.
(cherry picked from commit 629fb688f07b58ba50f5ce8f077da903dad2e073)
