+++ lib/ASTGen/Sources/ASTGen/ASTGen.swift
@@ -38,7 +38,7 @@ struct ASTGenVisitor: SyntaxTransformVisitor {
     fatalError("Use other overload.")
   }
 
-  public func visitAny(_ node: Syntax) -> ResultType {
+  public func visitAny(_ node: Syntax) -> UnsafeMutableRawPointer {
     fatalError("Not implemented.")
   }
   
@@ -59,254 +59,6 @@ struct ASTGenVisitor: SyntaxTransformVisitor {
     }
     
     return out
-  }
-  
-  public func visit(_ node: MemberDeclListItemSyntax) -> UnsafeMutableRawPointer {
-    visit(Syntax(node.decl))
-  }
-  
-  public func visit(_ node: StructDeclSyntax) -> UnsafeMutableRawPointer {
-    let loc = self.base.advanced(by: node.position.utf8Offset).raw
-    var nameText = node.identifier.text
-    let name = nameText.withUTF8 { buf in
-      return SwiftASTContext_getIdentifier(ctx, buf.baseAddress, buf.count)
-    }
-    
-    let out = StructDecl_create(ctx, loc, name, loc, declContext)
-    let oldDeclContext = declContext
-    declContext = out.declContext
-    defer { declContext = oldDeclContext }
-
-    node.members.members.map(self.visit).withBridgedArrayRef { ref in
-      NominalTypeDecl_setMembers(out.nominalDecl, ref)
-    }
-    
-    return out.decl
-  }
-  
-  public func visit(_ node: ClassDeclSyntax) -> UnsafeMutableRawPointer {
-    let loc = self.base.advanced(by: node.position.utf8Offset).raw
-    var nameText = node.identifier.text
-    let name = nameText.withUTF8 { buf in
-      return SwiftASTContext_getIdentifier(ctx, buf.baseAddress, buf.count)
-    }
-
-    let out = ClassDecl_create(ctx, loc, name, loc, declContext)
-    let oldDeclContext = declContext
-    declContext = out.declContext
-    defer { declContext = oldDeclContext }
-
-    node.members.members.map(self.visit).withBridgedArrayRef { ref in
-      NominalTypeDecl_setMembers(out.nominalDecl, ref)
-    }
-    
-    return out.decl
-  }
-  
-  public func visit(_ node: ClosureExprSyntax) -> UnsafeMutableRawPointer {
-    let statements = node.statements.map(self.visit)
-    let loc = self.base.advanced(by: node.position.utf8Offset).raw
-
-    let body = statements.withBridgedArrayRef { ref in
-      BraceStmt_create(ctx, loc, ref, loc)
-    }
-    
-    return ClosureExpr_create(ctx, body, declContext)
-  }
-
-  public func visit(_ node: FunctionCallExprSyntax) -> UnsafeMutableRawPointer {
-    // Transform the trailing closure into an argument.
-    if let trailingClosure = node.trailingClosure {
-      let tupleElement = TupleExprElementSyntax(label: nil, colon: nil, expression: ExprSyntax(trailingClosure), trailingComma: nil)
-      
-      return visit(node.addArgument(tupleElement).withTrailingClosure(nil))
-    }
-    
-    let args = visit(node.argumentList)
-    // TODO: hack
-    let callee = visit(node.calledExpression)
-
-    return SwiftFunctionCallExpr_create(self.ctx, callee, args)
-  }
-
-  public func visit(_ node: IdentifierExprSyntax) -> UnsafeMutableRawPointer {
-    let loc = self.base.advanced(by: node.position.utf8Offset).raw
-    
-    var text = node.identifier.text
-    let id = text.withUTF8 { buf in
-      return SwiftASTContext_getIdentifier(ctx, buf.baseAddress, buf.count)
-    }
-
-    return SwiftIdentifierExpr_create(ctx, id, loc)
-  }
-
-  public func visit(_ node: SimpleTypeIdentifierSyntax) -> UnsafeMutableRawPointer {
-    let loc = self.base.advanced(by: node.position.utf8Offset).raw
-    
-    var text = node.name.text
-    let id = text.withUTF8 { buf in
-      return SwiftASTContext_getIdentifier(ctx, buf.baseAddress, buf.count)
-    }
-
-    return SimpleIdentTypeRepr_create(ctx, loc, id)
-  }
-
-  public func visit(_ node: IdentifierPatternSyntax) -> UnsafeMutableRawPointer {
-    let loc = self.base.advanced(by: node.position.utf8Offset).raw
-    
-    var text = node.identifier.text
-    let id = text.withUTF8 { buf in
-      return SwiftASTContext_getIdentifier(ctx, buf.baseAddress, buf.count)
-    }
-
-    return SwiftIdentifierExpr_create(ctx, id, loc)
-  }
-  
-  public func visit(_ node: MemberAccessExprSyntax) -> UnsafeMutableRawPointer {
-    let loc = self.base.advanced(by: node.position.utf8Offset).raw
-    let base = visit(node.base!)
-    var nameText = node.name.text
-    let name = nameText.withUTF8 { buf in
-      return SwiftASTContext_getIdentifier(ctx, buf.baseAddress, buf.count)
-    }
-    
-    return UnresolvedDotExpr_create(ctx, base, loc, name, loc)
-  }
-  
-  public func visit(_ node: TupleExprElementSyntax) -> UnsafeMutableRawPointer {
-    visit(node.expression)
-  }
-
-  public func visit(_ node: TupleExprElementListSyntax) -> UnsafeMutableRawPointer {
-    let elements = node.map(self.visit)
-    
-    // TODO: find correct paren locs.
-    let lParenLoc = self.base.advanced(by: node.position.utf8Offset).raw
-    let rParenLoc = self.base.advanced(by: node.position.utf8Offset).raw
-    
-    return elements.withBridgedArrayRef { elementsRef in
-      SwiftTupleExpr_create(self.ctx, lParenLoc, elementsRef, rParenLoc)
-    }
-  }
-
-  public func visit(_ node: InitializerClauseSyntax) -> UnsafeMutableRawPointer {
-    visit(node.value)
-  }
-
-  public func visit(_ node: VariableDeclSyntax) -> UnsafeMutableRawPointer {
-    let pattern = visit(node.bindings.first!.pattern)
-    let initializer = visit(node.bindings.first!.initializer!)
-    
-    let loc = self.base.advanced(by: node.position.utf8Offset).raw
-    let isStateic = false // TODO: compute this
-    let isLet = node.letOrVarKeyword.tokenKind == .letKeyword
-
-    // TODO: don't drop "initializer" on the floor.
-    return SwiftVarDecl_create(ctx, pattern, loc, isStateic, isLet, declContext)
-  }
-  
-  public func visit(_ node: ConditionElementSyntax) -> UnsafeMutableRawPointer {
-    visit(node.condition)
-  }
-
-  public func visit(_ node: CodeBlockItemSyntax) -> UnsafeMutableRawPointer {
-    visit(node.item)
-  }
-
-  public func visit(_ node: CodeBlockSyntax) -> UnsafeMutableRawPointer {
-    let statements = node.statements.map(self.visit)
-    let loc = self.base.advanced(by: node.position.utf8Offset).raw
-
-    return statements.withBridgedArrayRef { ref in
-      BraceStmt_create(ctx, loc, ref, loc)
-    }
-  }
-
-  public func visit(_ node: FunctionParameterSyntax) -> UnsafeMutableRawPointer {
-    let loc = self.base.advanced(by: node.position.utf8Offset).raw
-    
-    let firstName: UnsafeMutableRawPointer?
-    let secondName: UnsafeMutableRawPointer?
-    
-    if let nodeFirstName = node.firstName {
-      var text = nodeFirstName.text
-      firstName = text.withUTF8 { buf in
-        SwiftASTContext_getIdentifier(ctx, buf.baseAddress, buf.count).raw
-      }
-    } else {
-      firstName = nil
-    }
-    
-    if let nodeSecondName = node.secondName {
-      var text = nodeSecondName.text
-      secondName = text.withUTF8 { buf in
-        SwiftASTContext_getIdentifier(ctx, buf.baseAddress, buf.count).raw
-      }
-    } else {
-      secondName = nil
-    }
-    
-    return ParamDecl_create(ctx, loc, loc, firstName, loc, secondName, declContext)
-  }
-
-  public func visit(_ node: FunctionDeclSyntax) -> UnsafeMutableRawPointer {
-    let loc = self.base.advanced(by: node.position.utf8Offset).raw
-    
-    var nameText = node.identifier.text
-    let name = nameText.withUTF8 { buf in
-      return SwiftASTContext_getIdentifier(ctx, buf.baseAddress, buf.count)
-    }
-    
-    let body: UnsafeMutableRawPointer?
-    if let nodeBody = node.body {
-      body = visit(nodeBody)
-    } else {
-      body = nil
-    }
-
-    let returnType: UnsafeMutableRawPointer?
-    if let output = node.signature.output {
-      returnType = visit(output.returnType)
-    } else {
-      returnType = nil
-    }
-    
-    let params = node.signature.input.parameterList.map { visit($0) }
-    return params.withBridgedArrayRef { ref in
-      FuncDecl_create(ctx, loc, false, loc, name, loc, false, nil, false, nil, loc, ref, loc, body, returnType, declContext)
-    }
-  }
-
-  public func visit(_ node: IfStmtSyntax) -> UnsafeMutableRawPointer {
-    let conditions = node.conditions.map(self.visit)
-    assert(conditions.count == 1) // TODO: handle multiple conditions.
-    
-    let body = visit(node.body)
-    let loc = self.base.advanced(by: node.position.utf8Offset).raw
-    
-    if let elseBody = node.elseBody, node.elseKeyword != nil {
-      return IfStmt_create(ctx, loc, conditions.first!, body, loc, visit(elseBody))
-    }
-    
-    return IfStmt_create(ctx, loc, conditions.first!, body, nil, nil)
-  }
-
-  public func visit(_ node: StringLiteralExprSyntax) -> UnsafeMutableRawPointer {
-    let loc = self.base.advanced(by: node.position.utf8Offset).raw
-    var segment = node.segments.first!.as(StringSegmentSyntax.self)!.content.text
-    return segment.withUTF8 { buf in
-      let id = SwiftASTContext_getIdentifier(ctx, buf.baseAddress, buf.count)
-      return SwiftStringLiteralExpr_create(ctx, id, buf.count, loc)
-    }
-  }
-
-  public func visit(_ node: IntegerLiteralExprSyntax) -> UnsafeMutableRawPointer {
-    let loc = self.base.advanced(by: node.position.utf8Offset).raw
-    var segment = node.digits.text
-    return segment.withUTF8 { buf in
-      let id = SwiftASTContext_getIdentifier(ctx, buf.baseAddress, buf.count)
-      return SwiftIntegerLiteralExpr_create(ctx, id, buf.count, loc)
-    }
   }
 }
